my order fix

+++ src/components/ManageOrder/ManageOrder.js
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from 'react';
 import { Link } from 'react-router-dom';
-import './ManageOrder.css'
+import './ManageOrder.css';
 
 
 const ManageOrder = () => {
@@ -11,53 +11,48 @@ const ManageOrder = () => {
         fetch('http://localhost:5000/manageOrders')
         .then(res=>res.json())
         .then(data=>setTours(data))
-    } ,[isDeleted])
+    } ,[])
 
     const handleDelete= (id)=>{
-        console.log(id);
-        fetch(`http://localhost:5000/deleteProduct/${id}`,{
-            method: "DELETE",
-            headers: {"content-type" : "application/json" },
-        })
-        .then(res=>res.json())
-        .then(result=> {
-            if(result.deleteCount>0){
-                setIsDeleted(true);
-            }
-            else{
-                setIsDeleted(false);
-            }
-        })
+        console.log(id)
+        const proceed = window.confirm("Are you sure , you want to delete?");
+        if(proceed){
+            fetch(`http://localhost:5000/deleteProduct/${id}`,{
+                method: "DELETE"
+            })
+            .then(res=>res.json())
+            .then(result=> {
+                if(result.deleteCount){
+                    alert("Deleted Successfully");
+                    const remaining = tours.filter(tour=>tour?._id !==id);
+                    setTours(remaining);
+                }
+            })
+        }
+
         console.log(id)
 
     }
     return (
-        <div >
-            <h2 className="my-5">Here tours are manages.....</h2>
+        <div className="container py-5 mb-0" >
+            <h2 className="my-5">Your All Orders Here.....</h2>
+
+            <div className="row bg-primary p-1">
+                <h5 className="col">Product id</h5>
+                <h5 className="col">Product name</h5>
+                <h5 className="col">User-email</h5>
+                <h5 className="col">Product price</h5>
+                <h5 className="col">Action</h5>
+                </div>
+
             {
-                tours.map((tour, index)=><div key={tour._id}>
-            <div className="d-flex justify-content-around">
-                
-                <div className="product-id">
-                    <h5 className="heading">Product id</h5>
-                    <p className="content">{tour._id}</p>
-                </div>
-                <div className="product-name">
-                    <h5 className="heading">Product name</h5>
-                    <p className="content">{tour.example}</p>
-                </div>
-                <div className="user-email">
-                    <h5 className="heading">User-email</h5>
-                    <p className="content">{tour.email}</p>
-                </div>
-                <div className="product-price">
-                    <h5 className="heading">Product price</h5>
-                    <p className="content">$ {tour.price}</p>
-                </div>
-                <div className="delete-button">
-                    <h5 className="heading">Action</h5>
-                    <button className="content" onClick={ ()=>handleDelete(tour._id)}> <i class="fas fa-trash-alt"></i> Cencle</button>
-                </div>
+                tours.map((tour)=><div key={tour?._id}>
+            <div className="row test-start bg-info ">
+            <h5 className="col">{tour.example}</h5>
+            <h5 className="col">{tour.email}</h5>
+            <h5 className="col">$ {tour.price}</h5>
+            <button className="col-lg-1 mb-2" onClick={ ()=>handleDelete(tour?._id)}> <i class="fas fa-trash-alt"></i> Cencle</button>
+            
             </div>
                     
             </div>)
